requirements.txt
langchain-openai>=0.0.2
langchain-core>=0.1.4
streamlit>=1.29.0
python-dotenv>=1.0.0